requirements.txt
loguru==0.7.2
requests~=2.32.3
nicegui==2.5.0
ruff==0.3.5
langgraph==0.2.39
pydantic~=2.8.2
python-dotenv==1.0.1
langchain==0.3.6
langchain-core~=0.3.12
matplotlib~=3.9.2
langchain_community==0.3.4
langchain-google-genai~=2.0.4
pdf2image~=1.17.0
pypdf~=5.1.0
pillow~=11.0.0
google-generativeai~=0.8.3
grpcio==1.60.1
langchain_chroma~=0.1.4
